requirements.txt
click==8.0.4
Flask==2.0.3
Flask-Markdown==0.3
importlib-metadata==4.11.2
itsdangerous==2.1.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
python-dotenv==0.19.2
Werkzeug==2.0.3
zipp==3.7.0
